requirements.txt
pyyaml>=6.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
keybert>=0.8.0
scikit-learn>=1.3.0
